requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cssselect2==0.8.0
Django==5.2
django-bootstrap4==25.1
django-crispy-forms==2.4
django-widget-tweaks==1.5.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.57.0
frozenlist==1.6.0
greenlet==3.2.1
gunicorn==23.0.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Markdown==3.5.1
MarkupSafe==3.0.2
multidict==6.4.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pdfkit==1.0.0
pillow==11.2.1
propcache==0.3.1
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
PySocks==1.7.1
python-dotenv==1.1.0
python-http-client==3.3.7
requests==2.32.3
selenium==4.15.2
sendgrid==6.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.40
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
tinycss2==1.4.0
tinyhtml5==2.0.0
trio==0.30.0
trio-websocket==0.12.2
twilio==9.6.0
typing_extensions==4.13.2
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.0
zopfli==0.2.3.post1
dj-database-url
pillow
reportlab
svglib
django
dj-database-url
python-dotenv
dj-database-url
Django==5.2
django-bootstrap4==25.1
django-crispy-forms==2.4
django-widget-tweaks==1.5.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
reportlab
requests==2.32.3
twilio==9.6.0
